refactor: promote custom open iddict application store to separate file
